Fix for drow crash on some binaries
The binaries that were used to repro this crash contain a RX segment with
no sections in it. This commit adds a check to ensure that the last
section is found in the RX segment. If it is not, it continues parsing
program headers in attempt to find a second RX segment.